데미지 주는 타이밍 변경(데이터 참조)

+++ Source/Core/Core/CoreTimeManager.h
@@ -3,7 +3,7 @@
 #define CORE_TIME_MANAGER GET_INSTANCE(CoreTimeManager)
 
 using TIME_POINT = std::chrono::system_clock::time_point;
-using TIME_VALUE = long long;
+using TIME_VALUE = int64_t;
 
 static constexpr TIME_VALUE SEC = 1000;
 static constexpr float_t SEC_FLOAT = 1000.0f;
